ip-by-geo/app/requirements.txt
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
maxminddb==1.4.1
maxminddb-geolite2==2018.703
python-geoip==1.2
pytz==2019.1
Werkzeug==0.15.4
